+++ core/components/organisms/alertService/alertService.tsx
@@ -1,8 +1,21 @@
 import * as React from 'react';
 import ReactDOM from 'react-dom';
+import { pubSub } from '../../../utils/alertServiceHelper';
 import AlertContainer, { AlertServiceToastProps } from './alertContainer';
-import { pubSub } from '../testHelper';
-
+const defaultConf = {
+  dismissIn: 3000,
+  autoHiderBar: {
+    style: {
+      height: '3px',
+      borderRadius: '5px',
+      backgroundColor: `rgb(0,0,0,35%)`
+    }
+  },
+  position: 'left',
+  transitionDelay: 800,
+  appearance: 'alert',
+  title: 'Some error occurred'
+};
 export interface AlertServiceConfig {
   dismissIn: number;
   position: string;
@@ -19,32 +32,25 @@ export interface PubSubServiceProps {
 export class AlertService {
   elem: HTMLElement;
   pubSubService: PubSubServiceProps;
-
-  constructor(
-    config = {
-      dismissIn: 3000,
-      autoHiderBar: {
-        style: {
-          height: '3px',
-          borderRadius: '5px',
-          backgroundColor: `rgb(0,0,0,35%)`
-        }
-      },
-      position: 'left',
-      transitionDelay: 800,
-      appearance: 'alert',
-      title: 'Some error occurred'
-    }
-  ) {
+  config: AlertServiceConfig;
+  constructor(config = {}) {
     this.elem = document.createElement('div');
     this.pubSubService = pubSub();
-    this.renderAlert(config);
+    this.config = { ...defaultConf, ...config };
+    this.renderAlert(this.config);
   }
 
   renderAlert = (config: AlertServiceConfig) => {
+    let fullConf = { ...this.config, ...config };
+    this.removeAlertService();
     this.elem.setAttribute('id', 'alertService-container');
     document.body.appendChild(this.elem);
-    ReactDOM.render(<AlertContainer pubSubService={this.pubSubService} defaultConfig={config} />, this.elem);
+    ReactDOM.render(<AlertContainer pubSubService={this.pubSubService} defaultConfig={fullConf} />, this.elem);
+  };
+
+  removeAlertService = () => {
+    const removed = ReactDOM.unmountComponentAtNode(this.elem);
+    return removed;
   };
 
   remove = (toastId: string) => this.pubSubService.publish('remove-toast', toastId);
