+++ core/index.tsx
@@ -79,6 +79,6 @@ export { Navigation } from './components/organisms/navigation';
 export { PageHeader } from './components/organisms/pageHeader';
 export { FileList } from './components/molecules/fileList';
 export { VerificationCodeInput } from './components/molecules/verificationCodeInput';
-export { AlertService }  from './utils/alertService';
+export { AlertService }  from './components/organisms/alertService';
 // @ts-ignore
 export { version } from '../package.json';
