+++ core/utils/index.tsx
@@ -1,4 +1,3 @@
 export * as css from './css';
 export * as validators from './validators';
 export * as masks from './masks';
-export * as AlertService from './alertService';
